Refactor code for pyproject
