+++ src/data_loader/dk_fullnames_to_shortnames.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 """
 Mapping of the full Names of the
 Desikan & Killiany regions to their abbreviations.
