+++ src/data_loader/ratio_exc_to_inh.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import pandas as pd
 
 __all__ = [
