+++ src/data_loader/synapse_cellbody_probability.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 """
 Provides the binzegger & mohan data.
