+++ src/data_loader/voneconomokoskinas.py
@@ -2,7 +2,6 @@
 
 import numpy as np
 import pandas as pd
-from scipy.io import loadmat
 
 __all__ = [
     "VonEconomoKoskinas",
