+++ src/figures/figure_appendix_rasterplot.py
