+++ src/figures/figure_firing_rate_distribution.py
