+++ src/figures/figure_spike_statistics.py
