+++ src/figures/figure_temporal_hierarchy.py
