+++ src/helpers/lognormal.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import numpy as np
 
 __all__ = [
