+++ src/network.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-import os
-import yaml
+
 import numpy as np
 import pandas as pd
-from copy import deepcopy
-from dicthash import dicthash
+
+# visible for notebook
+from .params.default_net_params import params as default_params
 
 __all__ = [
     "Network",
@@ -472,7 +472,7 @@ class Network():
 
         Returns
         -------
-        dc_drive : Series
+        dc_drive : pd.Series
             Vector of DC drives
         """
         dc_drive = pd.Series(
@@ -688,6 +688,8 @@ class Network():
         hash : str
             Hash for the network
         """
+        from dicthash import dicthash
+
         elem_net = self._elementarify_dict(self.net)
         hash_ = dicthash.generate_hash_from_dict(elem_net)
         return hash_
@@ -709,6 +711,9 @@ class Network():
         base_folder : string
             Path to base output folder
         """
+        import os
+        import yaml
+
         hash_ = self.getHash()
         out_folder = os.path.join(base_folder, hash_)
         try:
@@ -762,6 +767,8 @@ class Network():
         d : dict
             Modified copy of the dict
         """
+        from copy import deepcopy
+
         r = {}
         for key, val in d.items():
             if isinstance(val, (np.float32, np.float64)):
@@ -890,6 +897,10 @@ def networkDictFromDump(dump_folder):
     net_dit : dict
         Full network dictionary
     """
+    import os
+    import yaml
+    import pandas as pd
+
     # Read net.yaml
     fn = os.path.join(dump_folder, 'net.yaml')
     with open(fn, 'r') as net_file:
@@ -916,5 +927,3 @@ def networkDictFromDump(dump_folder):
         if fext == '.pkl':
             net_dict[fn] = pd.read_pickle(os.path.join(dump_folder, file))
     return net_dict
-
-def __dir__(): return sorted(__all__)
