+++ src/simulation.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
+
 import nest
 import numpy as np
 import os
-from datetime import datetime
-import yaml
-from dicthash import dicthash
 
 from .helpers.lognormal import mu_sigma_lognorm
+
+# visible for notebook
+from .params.default_sim_params import params as default_params
 
 __all__ = [
     "Simulation",
@@ -438,6 +439,8 @@ class Simulation():
 
     def simulate(self):
         """ Simulates the model."""
+        from datetime import datetime
+
         print("{} Start simulating".format(datetime.now()))
         nest.Simulate(self.sim_dict['t_sim'])
         print("{} Simulation finished".format(datetime.now()))
@@ -451,6 +454,8 @@ class Simulation():
         hash : str
             Hash for the simulation
         """
+        from dicthash import dicthash
+
         hash = dicthash.generate_hash_from_dict(self.sim_dict)
         return hash
 
@@ -479,6 +484,8 @@ class Simulation():
         base_folder : string
             Path to base output folder
         """
+        import yaml
+
         hash = self.getHash()
         out_folder = os.path.join(base_folder, hash)
         try:
@@ -506,10 +513,11 @@ def simulationDictFromDump(dump_folder):
     sim_dict : dict
         Full simulation dictionary
     """
+    import os
+    import yaml
+
     # Read sim.yaml
     fn = os.path.join(dump_folder, 'sim.yaml')
     with open(fn, 'r') as sim_file:
         sim_dict = yaml.load(sim_file, Loader=yaml.Loader)
     return sim_dict
-
-def __dir__(): return sorted(__all__)
