+++ src/snakemake/snakemake_helpers.py
@@ -1,7 +1,8 @@
+# -*- coding: utf-8 -*-
+
 """
 Some helper functions for snakemake.
 """
-import subprocess
 
 __all__ = [
     "get_git_revision_hash",
@@ -9,6 +10,8 @@ __all__ = [
 ]
 
 def get_git_revision_hash():
+    import subprocess as subprocess
+
     git_hash = subprocess.check_output(
             ['git', 'rev-parse', 'HEAD']
             ).decode('ascii').strip()
