+++ src/snakemake/snakemake_simulation.py
@@ -9,12 +9,11 @@ import os
 import sys
 import importlib.util
 
-from .network import networkDictFromDump
-from .simulation import simulationDictFromDump
-from .analysis import Analysis
-from .helpers.snakemake import nested_dict_update, get_git_revision_hash
-from .default_ana_params import params as ana_params
-from .default_net_params import params as net_params
+from .snakemake_helpers import nested_dict_update, get_git_revision_hash
+from ..params.default_sim_params import params as sim_params
+from ..params.default_net_params import params as net_params
+from ..network import networkDictFromDump
+from ..simulation import Simulation
 
 # Load script from specified path sys.argv[1] (snakemake {input})
 # Does nothing else than `import path/to/script as exp` would do
@@ -26,7 +25,7 @@ spec.loader.exec_module(exp)
 
 # Update default parameters with parameters specified in exp
 nested_dict_update(net_params, exp.net_params)
-nested_dict_update(ana_params, exp.ana_params)
+nested_dict_update(sim_params, exp.sim_params)
 
 outpath = net_params['outpath']
 
@@ -34,34 +33,31 @@ outpath = net_params['outpath']
 with open(sys.argv[2], 'r') as f:
     net_hash = f.read()
 
-# Read simulation hash
-with open(sys.argv[3], 'r') as f:
-    sim_hash = f.read()
-
 # Read network dict
 net_folder = os.path.join(outpath, net_hash)
 net_dict = networkDictFromDump(net_folder)
 
-# Read simulation dict
-sim_folder = os.path.join(outpath, net_hash, sim_hash)
-sim_dict = simulationDictFromDump(sim_folder)
+# Create Simulation class, export it and calculate the hash
+sim = Simulation(sim_params, net_dict)
+sim.dump(net_folder)
+sim_hash = sim.getHash()
 
-# Create Analysis class and export it
-base_path = sys.argv[5]
-ana = Analysis(ana_params, net_dict, sim_dict, sim_folder, base_path)
-ana.dump(sim_folder)
 
-# Do the analysis
-ana.fullAnalysis()
+# Set output directory according to hashes, instantiate the Network
+# and run the simulation
+data_path = os.path.join(outpath, net_hash, sim_hash)
+num_threads = int(sys.argv[4])
+sim.setup(data_path, num_threads)
+sim.simulate()
 
-# Save the analysis hash
-hash_file = sys.argv[4]
+# Save the network hash
+hash_file = sys.argv[3]
 hash_fn = os.path.join(os.getcwd(), hash_file)
 with open(hash_fn, 'w') as f:
-    f.write(ana.getHash())
+    f.write(sim_hash)
 
 # Save the humam repository hash
-ana_folder = os.path.join(sim_folder, ana.getHash())
-git_fn = os.path.join(ana_folder, "git_hash.txt")
+sim_folder = os.path.join(net_folder, sim_hash)
+git_fn = os.path.join(sim_folder, "git_hash.txt")
 with open(git_fn, 'w') as f:
     f.write(get_git_revision_hash())
